requirement.txt
allure-pytest==2.9.43
allure-python-commons==2.9.43
appdirs==1.4.4
atomicwrites==1.4.0
attrs==21.2.0
colorama==0.4.4
enum34==1.1.10
et-xmlfile==1.1.0
fs==2.4.13
helium==3.0.7
HTMLTestRunner-Python3==0.8.0
iniconfig==1.1.1
lxml==4.6.3
namedlist==1.8
nose==1.3.7
openpyxl==3.0.7
packaging==20.9
pluggy==0.13.1
py==1.10.0
pyparsing==2.4.7
pytest==6.2.4
pytest-failed-screenshot==1.0.2
pytz==2021.1
selenium==3.141.0
six==1.16.0
toml==0.10.2
urllib3==1.26.5
